litespi: move MetaSizes to separate file
